+++ src/client/sprites/Watermark.ts
@@ -1,0 +1,15 @@
+import { canvas } from '../canvas.ts';
+import Text, { type TextOptions } from './Text.ts';
+
+export default class Watermark extends Text {
+
+    constructor(options?: TextOptions) {
+        super(options);
+
+        this.content = options?.content ?? 'Made with TScratch';
+        this.x = options?.x ?? -canvas.width / 2 + 5,
+        this.y = options?.y ?? canvas.height / 2 - 5,
+        this.align = options?.align ?? 'left',
+        this.baseline = options?.baseline ?? 'top'
+    }
+}